Some changes
Add new controllers

+++ CaseApplication.EntityFramework/Repositories/CaseInventoryRepository.cs
@@ -52,9 +52,12 @@ namespace CaseApplication.EntityFramework.Repositories
         }
         public async Task<bool> CaseInventoryUpdate(CaseInventory caseInventory)
         {
-            CaseInventory searchCaseInventory = await GetCurrentCaseInventory(caseInventory.Id);
-
             using ApplicationDbContext _context = _contextFactory.CreateDbContext();
+
+            CaseInventory? searchCaseInventory = await _context.CaseInventory.FirstOrDefaultAsync(x => x.Id == caseInventory.Id);
+
+            if (searchCaseInventory is null) throw new Exception("There is no such case inventory in the database, " +
+                "review what data comes from the api");
 
             _context.Entry(searchCaseInventory).CurrentValues.SetValues(caseInventory);
             await _context.SaveChangesAsync();
@@ -64,9 +67,12 @@ namespace CaseApplication.EntityFramework.Repositories
 
         public async Task<bool> CaseInventoryDelete(Guid id)
         {
-            CaseInventory searchCaseInventory = await GetCurrentCaseInventory(id);
-
             using ApplicationDbContext _context = _contextFactory.CreateDbContext();
+
+            CaseInventory? searchCaseInventory = await _context.CaseInventory.FirstOrDefaultAsync(x => x.Id == id);
+
+            if (searchCaseInventory is null) throw new Exception("There is no such case inventory in the database, " +
+                "review what data comes from the api");
 
             _context.CaseInventory.Remove(searchCaseInventory);
             await _context.SaveChangesAsync();
